Initial LSM6DS33 IMU support (#37)
This commit adds support for the built-in LSM6DS33 6-axis IMU that is
installed on the Romi control board. The IMU is connected to the I2C bus
shared by the Pi and Romi 32U4.

This update does not currently expose the IMU data, although it does
include code to constantly update the IMU data as part of the sensor
update cycle.